feat: handle multiple social login providers
