+++ .env.example
@@ -14,6 +14,14 @@ DATABASE_URL="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST
 NEXTAUTH_SECRET=secret
 NEXTAUTH_URL=http://localhost:3000
 
+# Next Auth Github Provider
+GITHUB_CLIENT_ID=
+GITHUB_CLIENT_SECRET=
+
+# Next Auth Google Provider
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+
 # Api key to authenticate sync requests
 API_KEY=
